Move Frame from x86::addr to x86::page
